un-split StartIdentityVerification component
